requirements.txt
fastapi
uvicorn
python-docx
docx2pdf
requests
pydantic